Fixed Carousels issues
