+++ src/App.js
@@ -1,17 +1,8 @@
 import React, { Fragment } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { ChakraProvider } from "@chakra-ui/react";
-import Navbar from "./components/layouts/Navbar";
+import Home from "./components/Home/Home";
 import About from "./components/Pages/About";
-import Showcase from "./components/Showcase";
-import DiscoverHospitality from "./components/DiscoverHospitality";
-import Blog from "./components/Blog";
-import StudyProgrammes from "./components/StudyProgrammes";
-import OurStories from "./components/OurStories";
-import Partners from "./components/Partners";
-import Carousel from "./components/TestCar";
-import Footer from "./components/Footer";
-import FooterCopywright from "./components/FooterCopywright";
 
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
@@ -22,20 +13,9 @@ function App() {
       <Router>
         <Fragment>
           <Routes>
-            <Route exact path="/about" component={<About />} />
+            <Route exact path="/" element={<Home />} />
+            <Route exact path="/about" element={<About />} />
           </Routes>
-
-          <Navbar />
-          <Showcase />
-          <DiscoverHospitality />
-          <Blog />
-          <StudyProgrammes />
-          <OurStories />
-          <Partners />
-          <Carousel />
-
-          <Footer />
-          <FooterCopywright />
         </Fragment>
       </Router>
     </ChakraProvider>
