+++ src/components/Pages/Contact.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import Navbar from '../layouts/Navbar'
+
+const Contact = () => {
+  return (
+    <div>Contact</div>
+  )
+}
+
+export default Contact
